requirements.txt
PyYAML
mistralai
streamlit
